requirements.txt
langchain
tavily-python
llama-index
python-dotenv
newspaper3k
sentence-transformers
chromadb
cohere
streamlit